requirements.txt
customtkinter
pystray
pillow
pyinstaller
